Conflicts:
	repository/src/main/java/com/softserve/edu/config/JPATestingConfig.java
	repository/src/main/java/com/softserve/edu/repository/OrganizationRepository.java
	service/src/main/java/com/softserve/edu/service/admin/OrganizationService.java
	service/src/main/java/com/softserve/edu/service/admin/impl/OrganizationServiceImpl.java
	service/src/test/java/com/softserve/edu/service/admin/OrganizationsServiceTest.java
	service/src/test/java/com/softserve/edu/service/catalogue/LocalityServiceTest.java
	service/src/test/java/com/softserve/edu/service/catalogue/RegionServiceTest.java
	web/src/main/java/com/softserve/edu/controller/provider/ProviderApplicationController.java
	web/src/main/webapp/resources/app/admin/controllers/OrganizationEditModalController.js
	web/src/main/webapp/resources/app/admin/services/DataReceivingService.js
	web/src/main/webapp/resources/app/admin/views/modals/organization-add-modal.html
	web/src/main/webapp/resources/app/provider/services/VerificationServiceProvider.js
	web/src/main/webapp/resources/assets/i18n/welcome-ukr.json

+++ repository/src/main/java/com/softserve/edu/entity/verification/calibration/CalibrationTestData.java
@@ -24,8 +24,8 @@ public class CalibrationTestData {
     private Long id;
 
     private Double givenConsumption;
-    private Integer acceptableError;
-    private Integer volumeOfStandard;
+    private Long acceptableError;
+    private Long volumeOfStandard;
     private Double initialValue;
     private Double endValue;
     private Double volumeInDevice;
@@ -43,7 +43,7 @@ public class CalibrationTestData {
     private CalibrationTest calibrationTest;
 
     public CalibrationTestData(
-            Double givenConsumption, Integer acceptableError, Integer volumeOfStandard, Double initialValue,
+            Double givenConsumption, Long acceptableError, Long volumeOfStandard, Double initialValue,
             Double endValue, Double volumeInDevice, Double actualConsumption, String consumptionStatus,
             Double calculationError, CalibrationTestResult testResult, CalibrationTest calibrationTest
     ) {
